requirements.txt
python-redmine
python-ldap